% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_conversion.R
\name{abb_to_state}
\alias{abb_to_state}
\title{Convert Abbreviation to State}
\usage{
abb_to_state(abb)
}
\arguments{
\item{abb}{a two character capitalized abbreviation of a US state}
}
\value{
a capitalized name of the state associated with the given abbreviation
}
\description{
\code{abb_to_state()} and \code{state_to_abb()} convert state names into their
abbreviations, or abbreviations into names.
The list include District of Columbia, DC
}
\examples{
abb <- "NY"
state <- abb_to_state(abb) # New York

}
