% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_conversion.R
\name{is_state_abb}
\alias{is_state_abb}
\title{Is this a State Abbreviation?}
\usage{
is_state_abb(abb)
}
\arguments{
\item{abb}{a state abbreviation, capitalized}
}
\value{
TRUE if the input is a valid state abbreviation,
FALSE if it is not
}
\description{
Is this a State Abbreviation?
}
