% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_conversion.R
\name{is_state_name}
\alias{is_state_name}
\title{Is this a State Name?}
\usage{
is_state_name(state)
}
\arguments{
\item{state}{a state name, capitalized}
}
\value{
TRUE if the input is a valid state name,
FALSE if it is not
}
\description{
Is this a State Name?
}
\examples{
# Returns TRUE
is_state_name("Minnesota")

# Returns FALSE
is_state_name("Minesota")

# Returns FALSE, as it is not capitalized
is_state_name("minnesota")

}
