% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_food_access.R
\name{plot_state}
\alias{plot_state}
\title{Plot a state's food access}
\usage{
plot_state(
  data,
  state,
  feature = c("lila", "la"),
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  pal = c("YlOrRd", "Red-Yellow", "BluYl")
)
}
\arguments{
\item{data}{data frame - The foodatlas dataset}

\item{state}{character - The state to plot given in the format "NY" or "New York"}

\item{feature}{character - The feature to plot, either "lila" or "la"}

\item{title}{character - An optional title for the plot}

\item{subtitle}{character - An optional subtitle for the plot}

\item{caption}{character - An optional caption for the plot}

\item{pal}{character - The color pallete for the plot with options: "YlOrRd", "Red-Yellow" or "BluYl"}
}
\value{
A ggplot2 object
}
\description{
This function will generate a food access plot with county level detail.
Currently the only plot feature options are "lila" and "la" which refer to
"Low Income and Low Access" and "Low Access" repectively.
}
\examples{

plot_state(data = foodatlas, state = "NY", feature = "lila")

plot_state(foodatlas, "GA")

}
