% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_food_access.R
\name{plot_us}
\alias{plot_us}
\title{Plot food access across the US}
\usage{
plot_us(
  data,
  include_counties = FALSE,
  feature = c("lila", "la"),
  sensitivity = c("low", "medium", "high"),
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  pal = c("YlOrRd", "Red-Yellow", "BluYl")
)
}
\arguments{
\item{data}{data frame - The foodatlas dataset}

\item{include_counties}{bool - Plot just state data, or include counties as well}

\item{feature}{character - The feature to plot, either "lila" or "la"}

\item{sensitivity}{character - How strict should the plot be about what constitutes a low income or low access area? "low", "medium" or "high"}

\item{title}{character - An optional title for the plot}

\item{subtitle}{character - An optional subtitle for the plot}

\item{caption}{character - An optional caption for the plot}

\item{pal}{character - The color pallete for the plot with options: "YlOrRd", "Red-Yellow" or "BluYl"}
}
\value{
A ggplot2 object
}
\description{
Plot food access across the US
}
\examples{

plot_us(data = foodatlas, feature = "la")

plot_us(foodatlas, include_counties = "TRUE", feature = "lila", sensitivity = "medium",
        title = "Share of tracts per state with low income and low access to groceries")

}
