% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_conversion.R
\name{state_to_abb}
\alias{state_to_abb}
\title{Convert State To Abbreviation}
\usage{
state_to_abb(state)
}
\arguments{
\item{state}{a state name, capitalized}
}
\value{
a two character capitalized abbreviation of that state
}
\description{
\code{state_to_abb()} and \code{abb_to_state()} convert state names into their
abbreviations, or abbreviations into names.
The list include District of Columbia, DC
}
\examples{
state <- "New York"
abb <- state_to_abb(state) # "NY"

abb <- "AZ"
state <- abb_to_state(abb) # "Arizona"

}
